实现 大数相加

+++ calculator.hpp
@@ -3,8 +3,8 @@
 
 std::string Add(std::string n1, std::string n2)
 {
-    int num1 = std::stoi(n1);
-    int num2 = std::stoi(n2);
-    int result = num1 + num2;
+    long long num1 = std::stoll(n1);
+    long long num2 = std::stoll(n2);
+    long long result = num1 + num2;
     return std::to_string(result);
 }
